+++ packages/react-dom/src/server/ReactDOMServerNode.js
@@ -5,7 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import '../shared/ReactDOMInjection';
 import ReactVersion from 'shared/ReactVersion';
 
 import {renderToString, renderToStaticMarkup} from './ReactDOMStringRenderer';
